+++ pkg/factory/factory_test.go
@@ -64,7 +64,6 @@ var _ = Describe("Factory", func() {
 				"specs/in-progress",
 				"specs/completed",
 				false,
-				false,
 			)
 			Expect(processor).NotTo(BeNil())
 		})